Started working on process to send notifications.
The database will now store a column for an array of tokens. The idea currently (to try to get a push notification working) is to click a button and when that button is clicked it will just send a notification to whomever is logged in. The idea for the check in is that when the user selects a barber, the selected barber will be the only one to receive the notification.